Don't ignore .inrepo

+++ sh/clean-typescript-build.js
@@ -10,6 +10,7 @@ const files = glob.sync('packages/typescript/**/*.{js,d.ts,js.map}')
 
 for (const x of files) {
   if (/node_modules/.test(x)) continue
+  if (/\.inrepo/.test(x)) continue
   console.info('  --> Deleting', x)
   fs.unlink(x, error => error && console.error(`Failed to delete ${x}`, {error}))
 }
